fixup! Use an array for the position stack rather than an Erlang list
